+++ index.ts
@@ -18,7 +18,6 @@ export { Rating } from "./src/Rating";
 export { SegmentedControl } from "./src/SegmentedControl";
 export { ShimmerImage } from "./src/ShimmerImage";
 export { SkipLink } from "./src/SkipLink";
-export { Slider } from "./src/Slider";
 export { TextAnimation } from "./src/TextAnimation";
 export { TextReveal } from "./src/TextReveal";
 export { Tabs } from "./src/Tabs";