+++ src/Masonry/Masonry.spec.tsx
@@ -3,14 +3,14 @@ import { Masonry } from "./Masonry";
 
 describe("Masonry", () => {
   const children = [
-    <div>1</div>,
-    <div>2</div>,
-    <div>3</div>,
-    <div>4</div>,
-    <div>5</div>,
-    <div>6</div>,
-    <div>7</div>,
-    <div>8</div>,
+    <div key={1}>1</div>,
+    <div key={2}>2</div>,
+    <div key={3}>3</div>,
+    <div key={4}>4</div>,
+    <div key={5}>5</div>,
+    <div key={6}>6</div>,
+    <div key={7}>7</div>,
+    <div key={8}>8</div>
   ];
 
   it("render correct number of columns", () => {