+++ src/Slider/SlideTrack/SlideTrack.types.ts
@@ -1,3 +1,0 @@
-export interface SlideTrackPropsType {
-  handleTrackClick: (e: React.MouseEvent<HTMLSpanElement> | MouseEvent) => void;
-}